pulling out gradient checks as separate file
